refactor: 채팅 파일 3차 리팩토링 (#47)
