+++ src/features/chat/components/index.ts
@@ -7,12 +7,12 @@ export type { ChatInputMenuProps } from "@/features/chat/components/chat-detail/
 export type { ChatInputMenuAction } from "@/features/chat/types";
 export { ChatRoommateInviteSheet } from "@/features/chat/components/chat-detail/ChatRoommateInviteSheet";
 export type { ChatRoommateInviteSheetProps } from "@/features/chat/components/chat-detail/ChatRoommateInviteSheet";
-export { ChatRoommateInviteMessage } from "@/features/chat/components/ChatRoommateInviteMessage";
-export type { ChatRoommateInviteMessageProps } from "@/features/chat/components/ChatRoommateInviteMessage";
+export { ChatRoommateInviteMessage } from "@/features/chat/components/chat-detail/ChatRoommateInviteMessage";
+export type { ChatRoommateInviteMessageProps } from "@/features/chat/components/chat-detail/ChatRoommateInviteMessage";
 export { ChatRoommateConfirmedContent } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
 export type { ChatRoommateConfirmedContentProps } from "@/features/chat/components/chat-confirmed/ChatRoommateConfirmedContent";
-export { ChatRoommateRequestMessage } from "@/features/chat/components/ChatRoommateRequestMessage";
-export type { ChatRoommateRequestMessageProps } from "@/features/chat/components/ChatRoommateRequestMessage";
+export { ChatRoommateRequestMessage } from "@/features/chat/components/chat-detail/ChatRoommateRequestMessage";
+export type { ChatRoommateRequestMessageProps } from "@/features/chat/components/chat-detail/ChatRoommateRequestMessage";
 export { ChatMatchCard } from "@/features/chat/components/chat-detail/ChatMatchCard";
 export type { ChatMatchCardProps } from "@/features/chat/components/chat-detail/ChatMatchCard";
 export { ChatListItem } from "@/features/chat/components/chat-list/ChatListItem";
